fixed dublicate key error

+++ postgres/start.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS Categories (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name VARCHAR(500) NOT NULL,
     is_top_level BOOLEAN NOT NULL,
     parent_id INTEGER
 );
 CREATE TABLE IF NOT EXISTS Listings (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     title TEXT NOT NULL,
     data JSONB,
     kurzbeschreibung TEXT NOT NULL,
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS Listings (
 	    REFERENCES Categories(id)
 );
 CREATE TABLE IF NOT EXISTS AuctionResults (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     price INTEGER NOT NULL
 );
